LCD Control
Added LCD control.  Currently, it just turns the LCD on and off over and
over again pretty much forever, but it shows that I manged to get
communication.  I know it's pretty ugly, and there's a lot wrong with
it, but at least it works.  I'll fix it in a later update, I promise.